+++ impl/src/test/java/jakarta/faces/component/UIDataTest.java
@@ -1,4 +1,5 @@
 /*
+ * Copyright (c) 2023 Contributors to Eclipse Foundation.
  * Copyright (c) 1997, 2020 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
@@ -91,7 +92,7 @@ public class UIDataTest {
 
         data.setId("data");
         // simple way. otherwise, we have to mock the renderkit and whatever.
-        Whitebox.setInternalState(data, "clientId", data.getId());
+        Whitebox.setInternalState(data.getUiComponentBaseImpl(), "clientId", data.getId());
 
         data.invokeOnComponent(context, "differentId", (contextInLambda, target) -> {
         });
